Vivek Kumbhar Update June 18

+++ ms_production/ms_production/report/machine_shop_production_summary_analysis/machine_shop_production_summary_analysis.py
@@ -114,7 +114,7 @@ def get_columns():
   		{
 			"label": _("Earned Min"),
 	        	"fieldname": "earned_min",
-	          	"fieldtype": "Int",
+	          	"fieldtype": "Float",
 	           	"width": 150
 	        },
